pyproject.toml
[tool.poetry]
name = "animal-blog"
version = "0.1.0"
description = ""
authors = ["studentq <studentq.work@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.30"}
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
